Add Docz for documentation
Currently only has a few example pages, but is functional.
One issue is that the <Props /> component requires the component have .jsx extension, so we need to convert a few files.
Also, we should split state logic (e.g. Context) from generic components, so only generic components are used in the component library.