Remove aria-haspopup for dropdown component
Aria-haspopup declare that the dropdown is a menu component, and that it is possible to navigate with the arrow keys
I considered using aria-haspopup with said arrow key-navigation, but believe users might not expect that behaviour.
Therefore, I am going with removing aria-haspopup, to have it announced as a summary and details-component instead.