Add wcag-as-json as an internal library, instead of as a node_modules dependency
The npm package was not updated to reflect the May, 2020 update to WCAG 2.1.
Since this project is based on WCAG 2.1, we need those changes, and as such had to pull it into the project until such time as the npm package is updated.
All code, including licence, is included in the libraries/wcag-as-json folder